+++ modules/apps/101-interchain-swap/types/market_test.go
@@ -166,7 +166,7 @@ func TestSingleDeposit(t *testing.T) {
 		&pool,
 	)
 
-	pool.PoolPrice = float32(amm.LpPrice())
+	pool.PoolPrice = amm.LpPrice()
 
 	newDeposit := &types.Coin{
 		Amount: types.NewInt(initialY),
